Added type-cast
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@816 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/aoa.c
@@ -863,7 +863,7 @@ void AOACheckReference( theCell)
   long found = FALSE;
 
 #ifdef RTLAZY
-  if( *theCell > 0){
+  if( (int) *theCell > 0){
 #else
   if ( *theCell ){
 #endif
